antalia_project/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
django-phone-field==1.8.1
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
PyJWT==2.1.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
